refactor: 로또 번호 타입 LottoNumber로 변경
